+++ lib/FlashMem/FlashMem.h
@@ -49,7 +49,7 @@
 namespace Flash
 {
     /** Address of Meta Data header in EEPROM. */
-    static const uint8_t NVM_METADATA_ADDRESS = 0;
+    static const uint16_t NVM_METADATA_ADDRESS = 0;
 
     /** Meta Data maximal length. */
     static const uint8_t NVM_METADATA_MAX_LENGTH = 2;
@@ -58,13 +58,13 @@ namespace Flash
     static const String NVM_METADATA_VALID = "UZ";
 
     /** Address of saved SSID in EEPROM. */
-    static const uint8_t NVM_SSID_ADDRESS = NVM_METADATA_ADDRESS + NVM_METADATA_MAX_LENGTH;
+    static const uint16_t NVM_SSID_ADDRESS = NVM_METADATA_ADDRESS + NVM_METADATA_MAX_LENGTH;
 
     /** SSID maximal length */
     static const uint8_t NVM_SSID_MAX_LENGTH = 32;
 
     /** Address of saved Password in EEPROM. */
-    static const uint8_t NVM_PASSWORD_ADDRESS = NVM_SSID_ADDRESS + NVM_SSID_MAX_LENGTH;
+    static const uint16_t NVM_PASSWORD_ADDRESS = NVM_SSID_ADDRESS + NVM_SSID_MAX_LENGTH;
 
     /** WPA-PSK Password laximal length */
     static const uint8_t NVM_PASSWORD_MAX_LENGTH = 63;
@@ -73,7 +73,7 @@ namespace Flash
     static const uint8_t CREDENTIALS_MAX_LENGTH = NVM_PASSWORD_ADDRESS + NVM_PASSWORD_MAX_LENGTH;
 
     /** Address of saved Groups in EEPROM. */
-    static const uint8_t NVM_GROUPS_ADDRESS = NVM_SSID_ADDRESS + CREDENTIALS_MAX_LENGTH;
+    static const uint16_t NVM_GROUPS_ADDRESS = NVM_SSID_ADDRESS + CREDENTIALS_MAX_LENGTH;
 
     /** Length of saved Groups in EEPROM. */
     static const uint8_t NVM_GROUPS_LENGTH = 1;
@@ -134,7 +134,7 @@ namespace Flash
      *  @param[in] maxLength Maximum Length of the String.
      *  @param[out] output Buffer to save the String to.
      */
-    void getString(const uint8_t &address, const uint8_t &maxLength, String &output);
+    void getString(const uint16_t &address, const uint8_t &maxLength, String &output);
 
     /**
      *  Saves a Null-terminated String in the EEPROM.
@@ -144,7 +144,7 @@ namespace Flash
      *  @param[in] input String to save in EEPROM.
      *  @return If string written in EEPROM, returns true. Otherwise false.
      */
-    bool setString(const uint8_t &address, const uint8_t &maxLength, const String &input);
+    bool setString(const uint16_t &address, const uint8_t &maxLength, const String &input);
 
     /**
      *  Retrieves an 8-bit Unsigned Integer from the EEPROM.
@@ -154,7 +154,7 @@ namespace Flash
      *  @return If 8-bit Unsigned Integer succesfully retrieved from EEPROM, returns true. 
      *          Otherwise false.
      */
-    bool getUInt8(const uint8_t &address, uint8_t &value);
+    bool getUInt8(const uint16_t &address, uint8_t &value);
 
     /**
      *  Saves an 8-bit Unsigned Integer in the EEPROM.
@@ -163,7 +163,7 @@ namespace Flash
      *  @param[in] value 8-bit Unsigned Integer to save in EEPROM.
      *  @return If 8-bit Unsigned Integer written in EEPROM, returns true. Otherwise false.
      */
-    bool setUInt8(const uint8_t &address, uint8_t value);
+    bool setUInt8(const uint16_t &address, uint8_t value);
 
 };
